add unhide eye to options drawer:
