+++ frontend/src/pages/TermPlanner/main.less
@@ -232,35 +232,11 @@ body {
   }
   .unhideEye {
     transition: 200ms ease all;
-    color: #d2d0d0;
-    &:hover {
-      color: #9254de;
-    }
   }
 
-  //   .yearLineContainer {
-  //   }
-  //   .yearLine {
-  //     // border-bottom: 1px solid #d9d9d9;
-  //     grid-column: 1/ 5;
-  //     background-color: #d9d9d9;
-  //     height: 1px;
-  //     margin: 2em;
-  //     cursor: pointer;
-  //     &:hover {
-  //       grid-column: 2 / 5;
-  //       .unhideEye {
-  //         display: block;
-  //       }
-  //     }
-  //   }
-  //   .yearLineWithSummer {
-  //     grid-column: 1 / 6;
-  //   }
-
-  //   .unhideEye {
-  //     display: none;
-  //   }
+  .optionsHeader {
+    display: flex;
+  }
 }
 
 .dark {
@@ -368,6 +344,13 @@ body {
 
   .clicked {
     background-color: #242728;
+  }
+  .eye {
+    color: #373a3a;
+  }
+
+  .unhideEye {
+    color: #353838;
   }
 }
 
@@ -477,5 +460,12 @@ body {
     &:hover {
       background-color: #cecece;
     }
+  }
+
+  .eye {
+    color: #d0c8db;
+  }
+  .unhideEye {
+    color: #303539;
   }
 }
